requirements-tests.txt
-r requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements-tests.txt.tmp -
#
apipkg==1.5
    # via execnet
appdirs==1.4.4
    # via black
attrs==20.3.0
    # via pytest
beautifulsoup4==4.9.3
    # via bs4
black==20.8b1
    # via -r -
bs4==0.0.1
    # via -r -
chardet==4.0.0
    # via diff-cover
click==7.1.2
    # via
    #   black
    #   pip-tools
coverage==5.5
    # via pytest-cov
diff-cover==5.0.1
    # via -r -
execnet==1.8.0
    # via pytest-xdist
factory-boy==3.2.0
    # via pytest-factoryboy
faker==8.1.0
    # via factory-boy
flake8==3.9.1
    # via pytest-flake8
inflect==5.3.0
    # via jinja2-pluralize
inflection==0.5.1
    # via pytest-factoryboy
iniconfig==1.1.1
    # via pytest
isort==5.8.0
    # via pytest-isort
jinja2-pluralize==0.3.0
    # via diff-cover
jinja2==2.11.3
    # via
    #   diff-cover
    #   jinja2-pluralize
markupsafe==1.1.1
    # via jinja2
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
packaging==20.9
    # via pytest
pathspec==0.8.1
    # via black
pep517==0.10.0
    # via pip-tools
pip-tools==6.1.0
    # via -r -
pluggy==0.13.1
    # via
    #   diff-cover
    #   pytest
py==1.10.0
    # via
    #   pytest
    #   pytest-forked
pycodestyle==2.7.0
    # via flake8
pyflakes==2.3.1
    # via flake8
pygments==2.8.1
    # via diff-cover
pyparsing==2.4.7
    # via packaging
pytest-cov==2.11.1
    # via -r -
pytest-django==4.2.0
    # via -r -
pytest-factoryboy==2.1.0
    # via -r -
pytest-flake8==1.0.7
    # via -r -
pytest-forked==1.3.0
    # via pytest-xdist
pytest-isort==1.3.0
    # via -r -
pytest-xdist==2.2.1
    # via -r -
pytest==6.2.3
    # via
    #   -r -
    #   pytest-cov
    #   pytest-django
    #   pytest-factoryboy
    #   pytest-flake8
    #   pytest-forked
    #   pytest-xdist
python-dateutil==2.8.1
    # via faker
regex==2021.4.4
    # via black
six==1.15.0
    # via python-dateutil
soupsieve==2.2.1
    # via beautifulsoup4
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via
    #   black
    #   pep517
    #   pytest
typed-ast==1.4.3
    # via black
typing-extensions==3.7.4.3
    # via black

# The following packages are considered to be unsafe in a requirements file:
# pip
